+++ python/pod5/src/tests/test_writer.py
@@ -19,11 +19,27 @@ class TestPod5Writer:
     def test_writer_random_reads(self, writer: p5.Writer, random_read: p5.Read) -> None:
         """Write some random single reads to a writer"""
 
-        writer.add_read_object(random_read)
+        writer.add_read(random_read)
 
     @pytest.mark.parametrize("random_read_pre_compressed", [1], indirect=True)
     def test_writer_random_reads_compressed(
         self, writer: p5.Writer, random_read_pre_compressed: p5.Read
     ) -> None:
         """Write some random single reads to a writer which are pre-compressed"""
-        writer.add_read_object(random_read_pre_compressed)
+        writer.add_read(random_read_pre_compressed)
+
+    @pytest.mark.parametrize("random_read", [1, 2, 3, 4], indirect=True)
+    def test_writer_random_reads_deprecated(
+        self, writer: p5.Writer, random_read: p5.Read
+    ) -> None:
+        """Write some random single reads to a writer"""
+        with pytest.deprecated_call():
+            writer.add_read_object(random_read)
+
+    @pytest.mark.parametrize("random_read_pre_compressed", [1], indirect=True)
+    def test_writer_random_reads_compressed_deprecated(
+        self, writer: p5.Writer, random_read_pre_compressed: p5.CompressedRead
+    ) -> None:
+        """Write some random single reads to a writer which are pre-compressed"""
+        with pytest.deprecated_call():
+            writer.add_read_object(random_read_pre_compressed)